refactor(pdf) Move stream/buffer handling to pdf transform
